requirements.txt
streamlit>=1.28.0
requests>=2.31.0
numpy>=1.24.0
hnswlib>=0.7.0
pymupdf>=1.23.0
sentence-transformers>=2.2.0
pytesseract>=0.3.10
Pillow>=10.0.0